requirements.txt
anyio==3.7.1
asgiref==3.7.2
attrs==23.2.0
beautifulsoup4==4.12.3
bs4==0.0.2
certifi==2024.2.2
cffi==1.15.1
charset-normalizer==3.3.2
Django==3.2.25
djangorestframework==3.15.1
exceptiongroup==1.2.1
h11==0.14.0
httpcore==0.17.3
httpx==0.24.1
idna==3.7
importlib-metadata==6.7.0
outcome==1.3.0.post0
pycparser==2.21
PySocks==1.7.1
python-telegram-bot==20.3
pytz==2024.1
requests==2.31.0
selenium==4.11.2
sniffio==1.3.1
sortedcontainers==2.4.0
soupsieve==2.4.1
sqlparse==0.4.4
telegram==0.0.1
trio==0.22.2
trio-websocket==0.11.1
typing_extensions==4.7.1
urllib3==2.0.7
wsproto==1.2.0
zipp==3.15.0
